refactor(config): separate base definition from override logic
